Dev scud anthrax beta

+++ pyenv.py
@@ -9,21 +9,35 @@ import time
 import subprocess
 import os
 import json
+import sys
 import numpy as np
 from shutil import copy2
 import shutil
 import signal
-from common.util import write_prep_action, ControlObject
+from common.util import write_prep_action, ControlObject, get_initial_obs
+from common.obs_parsing import is_valid_action
 from common.metrics import Stopwatch
 
 # Config variables
-jar_name = 'tower-defence-runner-1.1.2.jar'
-config_name = 'config.json'
+jar_name = 'tower-defence-runner-2.0.3.jar'
+config_name = 'game-runner-config.json'#'config.json'
 game_config_name = 'game-config.properties'
 wrapper_out_filename = 'wrapper_out.txt'
 state_name = 'state.json'
 bot_file_name = 'bot.json'
-per_step_reward_penalty = -10
+per_step_reward_penalty = -8.0
+
+binary_step_penalty = -0.009
+
+general_reward_scaling_factor = 0.5
+
+binary_win_reward = 1.0
+dense_win_reward = 500
+
+possible_reward_modes = ['dense', 'binary']
+reward_mode = possible_reward_modes[0]
+
+keep_log_folder_override = False
 
 class Env():
 
@@ -36,6 +50,10 @@ class Env():
         # setting up jar runner
         self.needs_reset = True
         self.pid = None
+        self.done = False
+        self.prev_obs = get_initial_obs(1)[0][0]
+        self.clock = Stopwatch()
+        self.step_num = 0
 
     def setup_directory(self):
         # creates the dirs responsible for this env, 
@@ -82,14 +100,6 @@ class Env():
         if self.needs_reset:
             self.reset()
 
-        if self.proc.poll() != None:
-            # env ended last step, so reset:
-            if self.debug:
-                print(">> PYENV ", self.name ," >>  Ended early")
-            cntrl_obj = ControlObject('EARLY')
-            tp = np.concatenate([np.asarray([cntrl_obj,]), np.asarray([cntrl_obj,])], axis=-1)
-            return tp, np.concatenate([np.asarray([cntrl_obj,]), np.asarray([cntrl_obj,])], axis=-1)
-
         #######################
         ## Debug messages
         if self.debug:
@@ -98,7 +108,7 @@ class Env():
 
             with open(os.path.join(self.refbot_path, 'mylog.txt'), 'a') as f:
                 f.write(str(time.time()) + "\t-->Wanting to do op:!!!\t" + str(ref_act) + '\n')
-
+        ep_info = {}
         #######################
         ## Writing actions
         x2, y2, build2 = ref_act
@@ -120,6 +130,16 @@ class Env():
             if self.debug:
                 print(">> pyenv {} >> writing 2 to file {}".format(self.refenv.name, self.refenv.in_file))
             f.write('2')
+
+        #######################
+        ## Checking if episode ended early
+        if self.proc.poll() != None and self.done == True:
+            # env ended last step, so reset:
+            if self.debug:
+                print(">> PYENV ", self.name ," >>  Ended early")
+            cntrl_obj = ControlObject('EARLY')
+            tp = np.concatenate([np.asarray([cntrl_obj,]), np.asarray([cntrl_obj,])], axis=-1)
+            return tp, np.concatenate([np.asarray([cntrl_obj,]), np.asarray([cntrl_obj,])], axis=-1), ep_info
 
         #######################
         ## Taking step
@@ -162,22 +182,72 @@ class Env():
             if should_load_obs == True and should_load_obs2 == True:
                 break
             
-            if stopw.deltaT() > 8:
+            if self.proc.poll() != None and self.done == False:
+                #ep ended early.
+                if self.debug:
+                    print("PYENV: >> GAME ENDING EARLY FOR THE FIRST TIME")
+                self.done = True
+
+                valid, reason = is_valid_action(action, self.prev_obs)
+                obs = self.load_state()
+                self.prev_obs = obs
+                
+                ep_info['n_steps'] = self.step_num
+
+                if valid == True:
+                    ep_info['valid'] = True
+                else:
+                    ep_info['valid'] = False
+
+                ref_obs = self.refenv.load_state()
+                if obs['players'][0]['playerType'] == 'A':
+                    a_hp = obs['players'][0]['health']
+                    b_hp = obs['players'][1]['health']
+                else:
+                    a_hp = obs['players'][1]['health']
+                    b_hp = obs['players'][0]['health']
+                k = np.asarray([obs,])
+                u = np.asarray([ref_obs,])
+                return_obs = np.concatenate([k, u], axis=-1)
+                if reward_mode == 'dense':
+                    win_reward = dense_win_reward
+                    lose_reward = -1 * dense_win_reward
+                else:
+                    win_reward = binary_win_reward
+                    lose_reward = -1 * binary_win_reward
+                if a_hp > b_hp:
+                    # player a wins
+                    ep_info['winner'] = 'A'
+                    return return_obs, np.concatenate([np.asarray([win_reward,]), np.asarray([lose_reward,])], axis=-1), ep_info
+                elif a_hp < b_hp:
+                    ep_info['winner'] = 'B'
+                    return return_obs, np.concatenate([np.asarray([lose_reward,]), np.asarray([win_reward,])], axis=-1), ep_info
+                else:
+                    ep_info['winner'] = 'TIE'
+                    return return_obs, np.concatenate([np.asarray([0.0,]), np.asarray([0.0,])], axis=-1), ep_info
+
+            if stopw.deltaT() > 3:
                 # we have waited more than 3s, game clearly ended
                 self.needs_reset = True
                 failure = True
-                #if self.debug:
-                print('pyenv: env ' + str(self.name) + ' with pid ' + str(self.pid) + ' needs reset. (', should_load_obs, ',',should_load_obs2, ')' , time.time())
+                print('pyenv: env ' + str(self.name) + ' with pid ' + str(self.pid) + ' encountered error. (', should_load_obs, ',',should_load_obs2, ')' , time.time())
                 break
 
             time.sleep(0.01)
         # TODO: possibly pre-parse obs here and derive a reward from it?
-
+        
         #########################
         ## Loading the obs if their jar's ended properly
         #ref_obs, _ = self.refenv.step(ref_act)
         if should_load_obs:
+            valid, reason = is_valid_action(action, self.prev_obs)
             obs = self.load_state()
+            self.prev_obs = obs
+
+            if valid == True:
+                ep_info['valid'] = True
+            else:
+                ep_info['valid'] = False
         if should_load_obs2:
             ref_obs = self.refenv.load_state()
 
@@ -190,22 +260,36 @@ class Env():
         if failure == True:
             cntrl_obj = ControlObject('FAILURE')
             tp = np.concatenate([np.asarray([cntrl_obj,]), np.asarray([cntrl_obj,])], axis=-1)
-            return tp, np.concatenate([np.asarray([cntrl_obj,]), np.asarray([cntrl_obj,])], axis=-1)
+            return tp, np.concatenate([np.asarray([cntrl_obj,]), np.asarray([cntrl_obj,])], axis=-1), ep_info
+
+        # print('-----A------------->', obs['players'][0]['health'])
+        # print('-----B------------->', obs['players'][1]['health'])
+        self.step_num += 1
 
         ########################
         ## Forming rewards and packaging the obs into a good numpy form
         if obs is not None:
             # Infer reward:
             #reward = float(obs['players'][0]['score']) - float(obs['players'][1]['score'])
-            curS = float(obs['players'][0]['score'])
+            curS = float(obs['players'][0]['score']) * general_reward_scaling_factor
             self.score_delta = curS - self.score
             reward = self.score_delta + per_step_reward_penalty
             self.score = curS
 
+        if ref_obs is not None:
+            curS2 = float(ref_obs['players'][0]['score']) * general_reward_scaling_factor
+            self.refenv.score_delta = curS2 - self.refenv.score
+            ref_reward = self.refenv.score_delta + per_step_reward_penalty
+            self.refenv.score = curS2
+
         k = np.asarray([obs,])
         u = np.asarray([ref_obs,])
         return_obs = np.concatenate([k, u], axis=-1)
-        return return_obs, np.concatenate([np.asarray([reward,]), np.asarray([0.0,])], axis=-1)
+        if reward_mode == 'dense':
+            return return_obs, np.concatenate([np.asarray([reward,]), np.asarray([ref_reward,])], axis=-1), ep_info
+        elif reward_mode == 'binary':
+            return return_obs, np.concatenate([np.asarray([binary_step_penalty,]), np.asarray([binary_step_penalty,])], axis=-1), ep_info
+
 
     def load_state(self):
         '''
@@ -214,7 +298,7 @@ class Env():
         while os.path.isfile(self.state_file) == False:
             if self.debug:
                print(">> PYENV >> waiting for state file  ", self.state_file, ' to appear')
-            time.sleep(0.02)
+            time.sleep(0.01)
 
         flag = False
         while flag == False:
@@ -224,7 +308,8 @@ class Env():
                 break
             except json.decoder.JSONDecodeError as e:
                 k = None
-                print(">> PYENV >> Failed to decode json state! Got error ", e)
+                if self.debug:
+                    print(">> PYENV >> Failed to decode json state! Got error ", e)
                 time.sleep(0.01)
 
         return k
@@ -238,19 +323,20 @@ class Env():
         return np.concatenate([x, y], axis=-1)
 
     def reset(self):
+        self.step_num = 0
         if self.debug:
             with open(os.path.join(self.run_path, 'mylog.txt'), 'a') as f:
                 f.write(str(time.time()) + "\t-->RESETTING!!!\n")
 
             with open(os.path.join(self.refbot_path, 'mylog.txt'), 'a') as f:
                 f.write(str(time.time()) + "\t-->RESETTING!!!\n")
-        time.sleep(0.01)
 
         if self.proc is not None:
             self.proc.terminate()
             self.proc.wait()
         self.needs_reset = False
-        time.sleep(0.07)
+        self.done = False
+        time.sleep(0.01)
         # trying to kill jar wrapper of this env
         pid_file = os.path.join(self.run_path, 'wrapper_pid.txt')
         if os.path.isfile(pid_file):
@@ -284,30 +370,51 @@ class Env():
         ## Trying to kill jar wrapper of ref env
         refpid_file = os.path.join(self.refbot_path, 'wrapper_pid.txt')
         if os.path.isfile(refpid_file):
-            with open(refpid_file, 'r') as f:
-                wrapper_pid2 = int(f.read())
-                if wrapper_pid2 == 0:
-                    return None
-                else:
+            flag = False
+            while flag == False:
+                with open(refpid_file, 'r') as f:
                     try:
-                        os.kill(wrapper_pid2, signal.SIGTERM)
-                    except (PermissionError, ProcessLookupError) as e:
-                        if self.debug:
-                            print(">> PYENV ", self.name, " >> Attempted to close refbot wrapper pid ", wrapper_pid2, " but got ERROR ", e)
+                        wrapper_pid2 = int(f.read())
+                    except ValueError:
+                        continue
+                    if wrapper_pid2 == 0:
+                        flag = True
+                        return None
+                    else:
+                        flag = True
+                        try:
+                            os.kill(wrapper_pid2, signal.SIGTERM)
+                        except (PermissionError, ProcessLookupError) as e:
+                            if self.debug:
+                                print(">> PYENV ", self.name, " >> Attempted to close refbot wrapper pid ", wrapper_pid2, " but got ERROR ", e)
         else:
             if self.debug:
                 print(">> PYENV >> Attempted to close refbot wrapper pid but the wrapper pid file was not found ")
-        time.sleep(0.07)
+        time.sleep(0.01)
+        
+        #######################
+        ## Flushing matchlogs folder if env alive for over 1h
+        if self.clock.deltaT() >= 1800:
+            print(">> PYENV {} >> Env alive for over half an hour, flushing (deleting) matchlogs folder".format(self.name))
+            self.cleanup()
+            self.clock.reset()
+            print("Cleand.")
 
         command = 'java -jar ' + os.path.join(self.run_path, jar_name)
 
+        if sys.platform == "win32":
+            she = False
+        else:
+            she = True
+
         if self.debug:
-            self.proc = subprocess.Popen(command, shell=True ,cwd=self.run_path)
+            self.proc = subprocess.Popen(command, shell=she , stdout=subprocess.PIPE, cwd=self.run_path)
             print("Opened process: ", str(command), " with pid ", self.proc.pid)
         else:
-            self.proc = subprocess.Popen(command, shell=True, stdout=subprocess.DEVNULL, cwd=self.run_path)
+            self.proc = subprocess.Popen(command, shell=she, stdout=subprocess.DEVNULL, cwd=self.run_path)
         
         self.pid = self.proc.pid
+        time.sleep(0.01)
 
         return True
 
@@ -347,13 +454,30 @@ class Env():
         else:
             print(">> PYENV >> Attempted to close wrapper pid but the wrapper pid file was not found ")
         time.sleep(0.1)
-        # pid_file2 = os.path.join(self.refenv.ref_path, 'wrapper_pid.txt')
-        # with open(pid_file2, 'r') as f:
-        #     wrapper_pid2 = int(f.read())
-        #     if wrapper_pid2 == 0:
-        #         return None
-        #     else:
-        #         os.kill(wrapper_pid2, signal.SIGTERM)
+
+        refpid_file = os.path.join(self.refbot_path, 'wrapper_pid.txt')
+        if os.path.isfile(refpid_file):
+            flag = False
+            while flag == False:
+                with open(refpid_file, 'r') as f:
+                    try:
+                        wrapper_pid2 = int(f.read())
+                    except ValueError:
+                        continue
+                    if wrapper_pid2 == 0:
+                        flag = True
+                        return None
+                    else:
+                        flag = True
+                        try:
+                            os.kill(wrapper_pid2, signal.SIGTERM)
+                        except (PermissionError, ProcessLookupError) as e:
+                            if self.debug:
+                                print(">> PYENV ", self.name, " >> Attempted to close refbot wrapper pid ", wrapper_pid2, " but got ERROR ", e)
+        else:
+            if self.debug:
+                print(">> PYENV >> Attempted to close refbot wrapper pid but the wrapper pid file was not found ")
+        time.sleep(0.1)
 
         self.pid = None
         return True
@@ -364,9 +488,13 @@ class Env():
         if self.debug:
             print("Removing folder: ", log_path)
         try:
-            shutil.rmtree(log_path)
+            if keep_log_folder_override == False:
+                shutil.rmtree(log_path)
+            else:
+                print(">> PYENV >> OVERRIDE - Keeping log files.")
+            time.sleep(0.1)
         except Exception:
-            pass
+            print(">> PYENV >> Exception occured while removing matchlogs folder")
 
 class RefEnv():
     def __init__(self, env, debug=False):
@@ -376,60 +504,11 @@ class RefEnv():
         self.in_file = os.path.join(self.ref_path, wrapper_out_filename)
         self.state_file = os.path.join(self.ref_path, state_name)
         self.bot_file = os.path.join(self.ref_path, bot_file_name)
+        self.score = 0
+        self.score_delta = 0
 
     def step(self, action):
         raise NotImplementedError
-        # x, y, build = action
-        # write_prep_action(x, y, build, path=self.ref_path, debug=self.debug)
-
-        # with open(os.path.join(self.ref_path, 'mylog.txt'), 'a') as f:
-        #     f.write(str(time.time()) + "\t-->Writing refbot action:!!!\t" + str(action) + '\n')
-        
-        # with open(self.in_file, 'w') as f:
-        #     # we want start of a new step
-        #     if self.debug:
-        #         print(">> pyenv {} >> writing 2 to file {}".format(self.name, self.in_file))
-        #     f.write('2')
-        # obs = None
-        # #reward = None
-        # stopw = Stopwatch()
-        # should_load_obs = False
-
-        # while True:
-        #     with open(self.in_file, 'r') as ff:
-        #         k = ff.read()
-        #         try:
-        #             k = int(k)
-        #         except ValueError:
-        #             continue
-        #         if k == 1:
-        #             #print("just wrote 0 to the ", self.out_file)
-        #             # a new turn has just been processed
-        #             should_load_obs = True
-        #             break
-            
-        #     if stopw.deltaT() > 6:
-        #         # we have waited more than 3s, game clearly ended
-        #         self.needs_reset = True
-        #         #if self.debug:
-        #         print('pyenv: something very bad happened in refbot ', self.name, '. Took more than 6s to respond...', time.time())
-        #         break
-
-        #     time.sleep(0.01)
-        # # TODO: possibly pre-parse obs here and derive a reward from it?
-        
-        # # if obs is not None:
-        # #     # Infer reward:
-        # #     #reward = float(obs['players'][0]['score']) - float(obs['players'][1]['score'])
-        # #     curS = float(obs['players'][0]['score'])
-        # #     reward = curS - self.score
-        # #     self.score = curS
-        # if should_load_obs:
-        #     obs = self.load_state()
-
-        # if obs is None:
-        #     print(">> PY_ENV >> REF OBS IS NONE. (", self.name, ")")
-
         # return obs, 0
 
     def load_state(self):
@@ -449,7 +528,8 @@ class RefEnv():
                 break
             except json.decoder.JSONDecodeError as e:
                 k = None
-                print(">> REF ENV >> Failed to decode json state! Got error ", e)
+                if self.debug:
+                    print(">> REF ENV >> Failed to decode json state! Got error ", e)
                 time.sleep(0.01)
 
         return k
